+++ multistream/decoder_osce_apply.go
@@ -140,7 +140,7 @@ func (d *streamState) runOSCELACEChannel(native []int16, mode streamOSCELACEMode
 			fc.GainsQ16[sf] = ctrl.GainsQ16[sf]
 			fc.PitchL[sf] = ctrl.PitchL[sf]
 		}
-		fc.SignalType = int32(ctrl.SignalType)
+		fc.SignalType = ctrl.SignalType
 		numBits := ctrl.NumBits
 		if numBits < 0 {
 			numBits = 0
@@ -151,7 +151,7 @@ func (d *streamState) runOSCELACEChannel(native []int16, mode streamOSCELACEMode
 			state.lacePeriods[:],
 			state.laceApplyIn16[:streamOSCELACEFrameSamples],
 			&fc,
-			int32(numBits),
+			numBits,
 		)
 	}
 	switch mode {
